Adds functionality to multiple buttons
Alright, this commit is kind of a mess, but currently
the add book button and submit button of the resulting
form have been given functionality. However, the submit
button does not fully work as it does not currently
hide the form upon submission. The actual submission
action of the submit button was taken away as it would
refresh the page causing data to be lost. More work is
needed to bring this button to full functionality.